pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "loto"
version = "0.1.0"
description = "Planning utilities for lock-out/tag-out"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic",
    "pydantic-settings",
    "PyYAML",
    "networkx",
    "jinja2",
    "reportlab",
    "qrcode",
    "requests",
    "structlog",
    "typer",
    "python-dotenv",
    "pandas",
    "PyPDF2",
    "weasyprint",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "hypothesis",
    "black",
    "isort",
    "ruff",
    "mypy",
    "responses",
    "pre-commit",
    "httpx",
    "fastapi",
    "types-PyYAML",
]

[tool.black]
line-length = 88
target-version = ["py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["loto"]

[tool.hatch.build.targets.sdist]
include = ["loto", "README.md"]
